requirements.txt
Django>=5.2,<5.3
psycopg[binary]==3.2.12
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
django-cors-headers==4.4.0
django-environ==0.11.2
